Change MovieSlider into functional component
